# api.yaml
...
openapi: 3.0.0
paths:
  "/user/generate-nonce":
    post:
      operationId: UserController_generateNonce
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/GenerateNonceDto"
      responses:
        '201':
          description: ''
      tags:
      - User
  "/user/verify-signature":
    post:
      operationId: UserController_verifySignature
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/VerifySignatureDto"
      responses:
        '201':
          description: ''
      tags:
      - User
  "/user/me":
    get:
      operationId: UserController_getMe
      parameters: []
      responses:
        '200':
          description: ''
      tags:
      - User
      security:
      - bearer: []
  "/user/user/{id}":
    get:
      operationId: UserController_getUser
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: number
      responses:
        '200':
          description: ''
      tags:
      - User
      security:
      - bearer: []
  "/user":
    patch:
      operationId: UserController_updateUser
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UpdateUserDto"
      responses:
        '200':
          description: ''
      tags:
      - User
      security:
      - bearer: []
  "/user/upload/avatar":
    post:
      operationId: UserController_uploadAvatar
      parameters: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: ''
      tags:
      - User
      security:
      - bearer: []
  "/nft":
    post:
      operationId: NftController_mint
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateNftDto"
      responses:
        '201':
          description: ''
      tags:
      - NFT
      security:
      - bearer: []
    get:
      operationId: NftController_getTransactionByAction
      parameters: []
      responses:
        '200':
          description: ''
      tags:
      - NFT
      security:
      - bearer: []
...
  "/offer":
    post:
      operationId: OfferController_mint
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/OfferDto"
      responses:
        '201':
          description: ''
      tags:
      - Offer
      security:
      - bearer: []
    get:
      operationId: OfferController_getTransactionByAction
      parameters: []
      responses:
        '200':
          description: ''
      tags:
      - Offer
      security:
      - bearer: []
    delete:
      operationId: OfferController_deleteOffer
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      responses:
        '200':
          description: ''
      tags:
      - Offer
      security:
      - bearer: []
  "/offer/{id}":
    patch:
      operationId: OfferController_updateOffer
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UpdateOfferDto"
      responses:
        '200':
          description: ''
      tags:
      - Offer
      security:
      - bearer: []
...
  "/whitelist":
    post:
      operationId: WhitelistController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateWhitelistDto"
      responses:
        '201':
          description: ''
      tags:
      - Whitelist
      security:
      - bearer: []
    get:
      operationId: WhitelistController_getTransactionByAction
      parameters: []
      responses:
        '200':
          description: ''
      tags:
      - Whitelist
      security:
      - bearer: []
...
  "/offerHistory":
    post:
      operationId: OfferHistoryController_mint
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateOfferHistoryDto"
      responses:
        '201':
          description: ''
      tags:
      - OfferHistory
      security:
      - bearer: []
    get:
      operationId: OfferHistoryController_getTransactionByAction
      parameters: []
      responses:
        '200':
          description: ''
      tags:
      - OfferHistory
      security:
      - bearer: []
info:
  title: Solarr API Documentation
  description: Solarr API Documentation
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    GenerateNonceDto:
      type: object
      properties:
        address:
          type: string
      required:
      - address
    VerifySignatureDto:
      type: object
      properties:
        address:
          type: string
        signature:
          type: string
      required:
      - address
      - signature
    UpdateUserDto:
      type: object
      properties:
        username:
          type: string
        avatar:
          type: string
      required:
      - username
      - avatar
    CreateNftDto:
      type: object
      properties:
        name:
          type: string
        tokenId:
          type: number
        state:
          type: number
        metadata:
          type: string
        collectionFloor:
          type: number
        bankValuation:
          type: number
        ownerId:
          type: number
        collectionId:
          type: number
      required:
      - name
      - tokenId
      - state
      - metadata
      - collectionFloor
      - bankValuation
      - ownerId
      - collectionId
...
    OfferDto:
      type: object
      properties:
        loanAmount:
          type: number
        period:
          type: number
        apr:
          type: number
        loanType:
          type: boolean
        borrowerId:
          type: number
        lenderId:
          type: number
        collectionAddress:
          type: string
        tokenId:
          type: number
      required:
      - loanAmount
      - period
      - apr
      - loanType
      - borrowerId
      - lenderId
      - collectionAddress
      - tokenId
    UpdateOfferDto:
      type: object
      properties:
        loanAmount:
          type: number
        period:
          type: number
        state:
          type: number
      required:
      - loanAmount
      - period
      - state
...
    CreateWhitelistDto:
      type: object
      properties:
        collectionAddress:
          type: string
        tokenId:
          type: number
      required:
      - collectionAddress
      - tokenId
...
    CreateOfferHistoryDto:
      type: object
      properties:
        action:
          type: number
        offerId:
          type: number
      required:
      - action
      - offerId